Refactoring ContactStorage - prepare for use different ContactStorage realizations
